requirements.txt
loguru
pathlib
python-dotenv
fastapi
uvicorn
emails
pydantic[email]
pymongo
dnspython